[CS0025] Tối ưu hoá hiệu suất List => IEnumerable

+++ SubjectCrawlService1/Crawlers/Interfaces/ISubjectCrawler.cs
@@ -7,6 +7,6 @@ namespace SubjectCrawlService1.Crawlers.Interfaces
     public interface ISubjectCrawler
     {
         Task<Subject> Crawl(string discipline, string keyword1);
-        Task<Subject> Crawl(ushort courseId);
+        Task<Subject> Crawl(int courseId);
     }
 }
